+++ src/MiloneCliCore/PlatformUnix.fs
@@ -177,6 +177,10 @@ rule link
   description = link $out
   command = $${CC:-gcc} $in -o $out ${LINK_FLAGS}
 
+rule archive
+  description = archive $out
+  command = ar r $out $in
+
 build $milone_o: cc $milone_c | $milone_h
 build $milone_platform_o: cc $milone_platform_c | $milone_h
 
@@ -191,7 +195,8 @@ build $milone_platform_o: cc $milone_platform_c | $milone_h
     let pic =
       match p.BinaryType with
       | BinaryType.Exe -> []
-      | BinaryType.SharedObj -> [ "-fPIC" ]
+      | BinaryType.SharedObj
+      | BinaryType.StaticLib -> [ "-fPIC" ]
 
     let debug =
       if p.CDebug then
@@ -213,7 +218,8 @@ build $milone_platform_o: cc $milone_platform_c | $milone_h
   let linkFlags =
     let binaryFlag =
       match p.BinaryType with
-      | BinaryType.Exe -> []
+      | BinaryType.Exe
+      | BinaryType.StaticLib -> []
       | BinaryType.SharedObj -> [ "-shared" ]
 
     let lFlags =
@@ -265,8 +271,16 @@ build $milone_platform_o: cc $milone_platform_c | $milone_h
       |> List.append (p.ObjList |> List.map Path.toString)
       |> S.concat " "
 
+    let linkRule =
+      match p.BinaryType with
+      | BinaryType.Exe
+      | BinaryType.SharedObj -> "link"
+      | BinaryType.StaticLib -> "archive"
+
     build
-    |> cons "build $exe_file: link $milone_o $milone_platform_o "
+    |> cons "build $exe_file: "
+    |> cons linkRule
+    |> cons " $milone_o $milone_platform_o "
     |> cons objFiles
     |> cons "\n"